Changed InputManager's general purpose.
Before this commit, the class was responsible for processing the inputted files and passing them to the program's OutputManager. Now, it's purpose is to process inputted files and to buffer them before the user decides what to do. It also holds a reference to the program's InputNotifier, an implementation of the observer pattern to facilitate communication between the program's UI components.