Changed output logic and Added output configurations (Basic implementation)
The user will be able to change the output's specifications to their liking. This may include the number of frames per line, the horizondal gap between images, the vertical gap and the output format.

Output logic was changed to avoid creating output files identical to ones already created. If a queued-to-process file is already in the output, it will be elegantly ignored.